# apps/studio-api/openapi.yaml
openapi: 3.1.0
info:
  title: Fastforge Studio API
  version: 0.1.0
  description: |
    One contract, two hosts. `fastforge-studio serve` implements it against a
    local checkout; the hosted service implements it against D1 and R2. The web
    client is identical in both cases, and only `GET /v1/capabilities` tells the
    two apart.

    Every response is enveloped: successes as `{ "data": … }`, failures as
    `{ "error": { "code", "message" } }`. Surfaces a deployment does not serve
    answer `404 CAPABILITY_UNAVAILABLE` rather than 501 — in local mode the
    hosted features do not exist at all, and navigation never links to them.
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:7391
    description: Local — fastforge-studio serve
  - url: https://api.fastforge.dev
    description: Hosted
...
paths:
...
  /v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      summary: Read one project
      operationId: getProject
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/Project" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Projects]
      summary: Rename a project
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProjectRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/Project" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Projects]
      summary: Remove a project
      description: |
        In local mode this only unregisters the directory. Nothing on disk is
        deleted — the checkout is the user's, not Studio's.
      operationId: deleteProject
      responses:
        "204": { description: Removed }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: The request is malformed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: No such resource, or the capability is not served here
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable and machine-readable; clients branch on this.
              example: PROJECT_NOT_FOUND
            message:
              type: string
              description: For a human reading a toast or a log line.
...
    Platform:
      type: string
      enum: [android, ios, macos, windows, linux, web]
    Project:
      type: object
      required: [id, name, platforms, hasFastforgeConfig, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        path:
          type: string
          description: Local checkout. Local mode only.
        repo:
          type: string
          description: Git remote. Hosted mode only.
        platforms:
          type: array
          items: { $ref: "#/components/schemas/Platform" }
        hasFastforgeConfig:
          type: boolean
          description: >-
            Whether `.fastforge/config.yaml` exists. Without it there are no
            stores to read, which is what the stores page's empty state says.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    UpdateProjectRequest:
      type: object
      properties:
        name: { type: string }
...